Renamed ImPlatform.h to ImagingPlatform.h
