+++ src/libImaging/ImagingPlatform.h
